requirements.txt
Flask
numpy
pillow
CORS
keras
scikit-image
tensorflow